Created index.jsx to ReactDOM.render to fix Jest error
