+++ client/src/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import App from './components/App.jsx';
+
+ReactDOM.render(<App />, document.getElementById('App'));
