+++ webpack.config.js
@@ -4,7 +4,7 @@ var DIST_DIR = path.join(__dirname, '/client/dist');
 var webpack = require('webpack');
 
 module.exports = {
-  entry: `${SRC_DIR}/app.jsx`,
+  entry: `${SRC_DIR}/index.jsx`,
   output: {
     filename: 'bundle.js',
     path: DIST_DIR
